removed debugging code
git-svn-id: svn://136.177.114.72/svn_GW/IPhreeqc/trunk@4272 1feff8c3-07ed-0310-ac33-dd36852eb9cd

+++ test/kinn.f
@@ -11,7 +11,7 @@ C      PROGRAM DRIVER
       CHARACTER(30) comp
       INTEGER n
       INTEGER len
-      INTEGER(KIND=4) id
+      INTEGER id
       REAL*8 dvalue
       
       id = CreateIPhreeqc()
